refactor: move BioCardList into GenericPagesComponents
