+++ src/components/GenericPagesComponents/BioCardList/index.ts
