+++ src/pages/about-us/leadership.tsx
@@ -1,14 +1,13 @@
 import { NextPage, GetStaticProps, GetStaticPropsResult } from "next";
 
-import CMSClient from "../../node-lib/cms";
-import AboutContactCard from "../../components/GenericPagesComponents/AboutContactCard";
-import AboutUsSummaryCard from "../../components/pages/AboutUs/AboutUsSummaryCard";
-import BioCardList from "../../components/BioCardList";
-import AboutIntroCard from "../../components/ArchivedComponents/AboutIntoCard/AboutIntroCard";
-import { getSeoProps } from "../../browser-lib/seo/getSeoProps";
-import { AboutLeadershipPage } from "../../common-lib/cms-types";
-import getPageProps from "../../node-lib/getPageProps";
-
+import CMSClient from "@/node-lib/cms";
+import AboutContactCard from "@/components/GenericPagesComponents/AboutContactCard";
+import AboutUsSummaryCard from "@/components/pages/AboutUs/AboutUsSummaryCard";
+import BioCardList from "@/components/GenericPagesComponents/BioCardList";
+import AboutIntroCard from "@/components/ArchivedComponents/AboutIntoCard/AboutIntroCard";
+import { getSeoProps } from "@/browser-lib/seo/getSeoProps";
+import { AboutLeadershipPage } from "@/common-lib/cms-types";
+import getPageProps from "@/node-lib/getPageProps";
 import MaxWidth from "@/components/SharedComponents/MaxWidth";
 import Layout from "@/components/SharedComponents/Layout";
 import { Heading } from "@/components/SharedComponents/Typography";
